Make max and min times floats

+++ src/main/java/tech/arhan/randomswap/RandomSwapDataStore.java
@@ -7,8 +7,8 @@ import net.minecraft.world.entity.player.Player;
 
 public class RandomSwapDataStore {
   private static final List<Player> players = new ArrayList<Player>();
-  private static int minTime = 0;
-  private static int maxTime = 10;
+  private static float minTime = 0;
+  private static float maxTime = 10;
   private static boolean showLostItem = true;
   private static boolean showGainedItem = true;
   private static boolean countdownStarted = false;
@@ -25,19 +25,19 @@ public class RandomSwapDataStore {
     players.clear();
   }
 
-  public static void setMinTime(int minutes) {
+  public static void setMinTime(float minutes) {
     minTime = minutes;
   }
 
-  public static int getMinTime() {
+  public static float getMinTime() {
     return minTime;
   }
 
-  public static void setMaxTime(int minutes) {
+  public static void setMaxTime(float minutes) {
     maxTime = minutes;
   }
 
-  public static int getMaxTime() {
+  public static float getMaxTime() {
     return maxTime;
   }
